refactor(departments): replace Popover with Sheet for member list, rename component
Sheet gives more vertical space for departments with many members.
